# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP_HW API",
        "version": "0.1"
    },
    "paths": {
        "/api/jobs": {
            "post": {
                "tags": [
                    "job"
                ],
                "summary": "Create job",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/job/{id}": {
            "get": {
                "tags": [
                    "job"
                ],
                "summary": "Get detail job",
                "description": "Get Job by Id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pet that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "maximum": 10,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "job"
                ],
                "summary": "Delete job",
                "description": "Delete Job by Id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of pet that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "maximum": 10,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "JobRequest": {
                "properties": {
                    "urls": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            }
        }
    }
}